requirements.txt
simple-image-download
PILLOW
tensorflow
numpy
opencv-python
